requirements.txt
jmcomic==2.5.32
nonebot==1.9.1
pikepdf==9.5.2
psutil==6.1.1
PyYAML==6.0.2
